BW64 major file type addition
c files for axml.c and chna.c chunks.
updates and additions to common.h, sndfile.c, sndfile.h.in , wavlike.h, wavlike.c
bw64_test.c is able to check import of a BW64 ADM encoded file, but the chunk ordering and chna and xml data does not appear to be in the expected positions in the header. Possibly an issue with header pointers updating incorrectly.